wounderful/requirements.txt
asgiref==3.2.7
six==1.14.0
sqlparse==0.3.1
Django==3.0.4
Jinja2==2.11.1
future==0.18.2
django-crispy-forms==1.9.2
dash==1.4.1
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
django-plotly-dash==1.0.1
plotly==4.2.1
pillow==7.2.0
scipy==1.5.2
DateTime==4.3
opencv-python==4.4.0.42
Keras==2.4.3
scikit-image==0.17.2
tensorflow==2.3.0
